batch/requirements.txt
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.134
botocore==1.12.134
bz2file==0.98
certifi==2019.3.9
chardet==3.0.4
Click==7.0
docutils==0.14
gensim==3.7.2
idna==2.8
itsdangerous==1.1.0
jmespath==0.9.4
MarkupSafe==1.1.1
mecab-python3==0.996.2
numpy==1.16.3
pycurl==7.43.0
pygobject==3.20.0
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.8.0
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
six==1.12.0
smart-open==1.8.2
soupsieve==1.9.1
unattended-upgrades==0.1
urllib3==1.24.2
Werkzeug==0.15.2
sqlalchemy
sqlalchemy_utils
pymysql
neologdn
mysqlclient